TVTBparam class objects

The TVTBparam class stores recurrent parameters of the TVTB package in a convenient format.

TVTBparam(
        genos, ranges = GRangesList(),
        aaf = "AAF", maf = "MAF", vep = "CSQ", bp = SerialParam(),
        svp = ScanVcfParam(which = reduce(unlist(ranges))))

Accessor methods

In the following code snippets x is a TVTBparam object.

genos(x), genos(x) <- value

Gets or sets the Genotypes object stored in the genos slot.

ranges(x), ranges(x) <- value

List of genomic ranges to group variants during analyses and plots.

ref(x), ref(x) <- value

Gets or sets the character vector that declares homozygote reference genotypes.

het(x), het(x) <- value

Gets or sets the character vector that declares heterozygote genotypes.

alt(x), alt(x) <- value

Gets or sets the character vector that declares homozygote alternate genotypes.

carrier(x)

Gets a character vectors of concatenated heterozygote and homozygote alternate genotypes. See also het and alt accessors.

aaf(x), aaf(x) <- value

Gets or sets the INFO key suffix used to store the alternate allele frequency (AAF).

maf(x), maf(x) <- value

Gets or sets the INFO key suffix used to store the minor allele frequency (MAF).

vep(x), maf(x) <- value

Gets or sets the INFO key suffix used to extract the VEP predictions.

bp(x), bp(x) <- value

Gets or sets the BiocParallel parameters.

suffix(x)

Gets a named character vector that declares individual suffixes used to store the data for each set of genotypes in the INFO field of the VCF object. Names of this vector are ref, het, alt, aaf, and maf.

svp(x), svp(x) <- value

Gets or sets the ScanVcfParam parameters.

Arguments

genos

A Genotypes object that declares the three sets of homozygote reference, heterozygote, and homozygote alternate genotypes, as well as the individual key suffix used to store data for each set of genotypes in the info slot of a VCF object. See also Details section.

ranges

A GRangesList of genomic regions. See svp argument. In the future, may be used to facet statistics and figures.

aaf

INFO key suffix used to store the alternate allele frequency (AAF).

maf

INFO key suffix used to store the minor allele frequency (MAF).

vep

INFO key suffix used to extract the VEP predictions. See svp argument.

bp

A BiocParallelParam object.

svp

A ScanVcfParam object. If none is supplied, the ScanVcfParam slot which is automatically set to reduce(unlist(ranges)).

Value

A TVTBparam object that contains recurrent parameters.

Details

For each suffix stored in the TVTBparam object, TVTB may store data in the VCF object under the INFO keys defined as follows:

suffix

Statistics across all samples in the ExpandedVCF (e.g. "MAF").

phenotype_level_suffix

Statistics across samples associated with a given level of a given phenotype (e.g. "gender_male_MAF").

Users are recommended to avoid using those INFO keys for other purposes.

Author

Kevin Rue-Albrecht

See also

Examples

# Constructors ----

grl <- GenomicRanges::GRangesList(GenomicRanges::GRanges(
    "15", IRanges::IRanges(48413170, 48434757, names = "SLC24A5")
    ))

tparam <- TVTBparam(Genotypes("0|0", c("0|1", "1|0"), "1|1"), ranges = grl)

# Accessors ----

## The Genotypes object stored in the genos slot of the TVTBparam object
## return by the genos() accessor.
genos(tparam)
#> Class: Genotypes
#>   @ref (hom. ref.): "REF" {0|0}
#>   @het (heter.)   : "HET" {0|1, 1|0}
#>   @alt (hom. alt.): "ALT" {1|1}

## Genomic ranges stored in the TVTBparam object returned by the ranges()
## accessor.
ranges(tparam)
#> GRangesList object of length 1:
#> [[1]]
#> GRanges object with 1 range and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>   SLC24A5       15 48413170-48434757      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

## Individual genotypes can be extracted with ref(), het(), alt() accessors.
ref(tparam)
#> [1] "0|0"
het(tparam)
#> [1] "0|1" "1|0"
alt(tparam)
#> [1] "1|1"

## Their individual INFO key suffixes can be extracted with suffix() applied to
## the above accessors.
suffix(tparam)
#>   ref   het   alt   aaf   maf 
#> "REF" "HET" "ALT" "AAF" "MAF" 
suffix(tparam)["ref"]
#>   ref 
#> "REF" 
suffix(tparam)["het"]
#>   het 
#> "HET" 
suffix(tparam)["alt"]
#>   alt 
#> "ALT" 
suffix(tparam)["aaf"]
#>   aaf 
#> "AAF" 
suffix(tparam)["maf"]
#>   maf 
#> "MAF" 

## Heterozygote, and alternate heterozygote genotypes are
## returned by the carrier() accessor.
carrier(tparam)
#> [1] "0|1" "1|0" "1|1"

## INFO key suffix of alternate/minor allele frequency returned by the aaf()
## and maf() accessors.
aaf(tparam)
#> [1] "AAF"
maf(tparam)
#> [1] "MAF"

## INFO key suffix of the VEP predictions returned by the vep() accessor.
vep(tparam)
#> [1] "CSQ"

## BiocParallel parameters
bp(tparam)
#> class: SerialParam
#>   bpisup: FALSE; bpnworkers: 1; bptasks: 0; bpjobname: BPJOB
#>   bplog: FALSE; bpthreshold: INFO; bpstopOnError: TRUE
#>   bpRNGseed: ; bptimeout: NA; bpprogressbar: FALSE
#>   bpexportglobals: FALSE; bpexportvariables: FALSE; bpforceGC: FALSE
#>   bpfallback: FALSE
#>   bplogdir: NA
#>   bpresultdir: NA

## ScanVcfParam parameters
svp(tparam)
#> class: ScanVcfParam 
#> vcfWhich: 1 elements
#> vcfFixed: character() [All] 
#> vcfInfo:  
#> vcfGeno:  
#> vcfSamples: