Add count of genotypes to INFO field

Adds the total occurences of a set of genotypes as an INFO field for each variant. All given genotypes are counted toward a single total (e.g. grand total of c("0/0", "0|0")), while other genotypes are silently ignored.

# S4 method for class 'ExpandedVCF'
addCountGenos(
        vcf, genos, key, description,
        samples = 1:ncol(vcf), force = FALSE)

Arguments

vcf

ExpandedVCF object.

genos

character vector of genotypes to count (toward a common unique total).

key

Name of the INFO field to create or update (character vector of length 1). See Details below.

description

character description of the INFO field to create or overwrite (character vector of length 1).

samples

integer, numeric or character vector indicating samples to consider in VariantAnnotation::geno(vcf). If not specified, all samples are considered.

force

If TRUE, the field header and data will be overwritten if present; If FALSE, an error is thrown if the field already exists.

Value

ExpandedVCF object including an additional INFO field stating the count of genotypes.

Author

Kevin Rue-Albrecht

Details

In all cases, the new INFO field is inserted after the last existing field. In other words, overwriting an existing INFO field is achieved by dropping it from the data and header of the info slot, and subsequently inserting the new data after the last remaining INFO field.

See also

Examples

# Example data ----

# VCF file
vcfFile <- system.file("extdata", "moderate.vcf", package = "TVTB")

# TVTB parameters
tparam <- TVTBparam(Genotypes(ref = "0|0", het = c("0|1", "1|0"), alt = "1|1"))

# Pre-process variants
vcf <- VariantAnnotation::readVcf(vcfFile, param = tparam)
vcf <- VariantAnnotation::expand(vcf, row.names = TRUE)

# Example usage ----

vcf <- addCountGenos(
    vcf, het(tparam),
    suffix(tparam)["het"],
    "Number of heterozygous genotypes")