Parse the CSQ column of a VCF object into a GRanges object

Parse the CSQ column in a VCF object returned from the Ensembl Variant Effect Predictor (VEP).

**This method was rescued following the deprecation of the package ensemblVEP in the Bioconductor release 3.20.**

# S4 method for class 'VCF'
parseCSQToGRanges(x, VCFRowID=character(),
    ..., info.key = "CSQ")

Arguments

x

A VCF object.

VCFRowID

A character vector of rownames from the original VCF. When provided, the result includes a metadata column named ‘VCFRowID’ which maps the result back to the row (variant) in the original VCF.

When VCFRowID is not provided no ‘VCFRowID’ column is included.

info.key

The name of the INFO key that VEP writes the consequences to in the output (default is CSQ). This should only be used if something other that CSQ was passed in the –vcf_info_field flag in the output options.

...

Arguments passed to other methods. Currently not used.

Details

-

When ensemblVEP returns a VCF object, the consequence data are returned unparsed in the 'CSQ' INFO column. parseCSQToGRanges parses these data into a GRanges object that is expanded to match the dimension of the 'CSQ' data. Because each variant can have multiple matches, the ranges in the GRanges are repeated.

If rownames from the original VCF are provided as VCFRowID a metadata column is included in the result that maps back to the row (variant) in the original VCF. This option is only applicable when the info.key field has data (is not empty).

If no info.key column is found the function returns the data in rowRanges().

Value

Returns a GRanges object with consequence data as the metadata columns. If no 'CSQ' column is found the GRanges from rowRanges() is returned.

Author

Valerie Obenchain, Kevin Rue-Albrecht

References

Ensembl VEP Home: http://uswest.ensembl.org/info/docs/tools/vep/index.html

Examples

  library(VariantAnnotation)
#> Loading required package: BiocGenerics
#> 
#> Attaching package: 'BiocGenerics'
#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from 'package:base':
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, table, tapply,
#>     union, unique, unsplit, which.max, which.min
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: 'MatrixGenerics'
#> The following objects are masked from 'package:matrixStats':
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomeInfoDb
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> 
#> Attaching package: 'S4Vectors'
#> The following object is masked from 'package:utils':
#> 
#>     findMatches
#> The following objects are masked from 'package:base':
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomicRanges
#> Loading required package: SummarizedExperiment
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: 'Biobase'
#> The following object is masked from 'package:MatrixGenerics':
#> 
#>     rowMedians
#> The following objects are masked from 'package:matrixStats':
#> 
#>     anyMissing, rowMedians
#> Loading required package: Rsamtools
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: 'Biostrings'
#> The following object is masked from 'package:base':
#> 
#>     strsplit
#> 
#> Attaching package: 'VariantAnnotation'
#> The following object is masked from 'package:base':
#> 
#>     tabulate
  file <- system.file("extdata", "moderate.vcf", package = "TVTB")
  vep <- readVcf(file)

  ## The returned 'CSQ' data are unparsed.
  info(vep)$CSQ
#> CharacterList of length 29
#> [[1]] C|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST0000...
#> [[2]] T|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST0000...
#> [[3]] A|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST0000...
#> [[4]] T|upstream_gene_variant|MODIFIER|SLC24A5|ENSG00000188467|Transcript|ENS...
#> [[5]] A|upstream_gene_variant|MODIFIER|SLC24A5|ENSG00000188467|Transcript|ENS...
#> [[6]] C|intron_variant|MODIFIER|SLC24A5|ENSG00000188467|Transcript|ENST000003...
#> [[7]] G|non_coding_transcript_exon_variant&non_coding_transcript_variant|MODI...
#> [[8]] G|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST0000...
#> [[9]] A|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST0000...
#> [[10]] G|missense_variant|MODERATE|SLC24A5|ENSG00000188467|Transcript|ENST000...
#> ...
#> <19 more elements>

  ## Parse into a GRanges and include the 'VCFRowID' column.
  vcf <- readVcf(file, "hg19")
  csq <- parseCSQToGRanges(vep, VCFRowID=rownames(vcf))
  csq[1:4]
#> GRanges object with 4 ranges and 65 metadata columns:
#>               seqnames    ranges strand |  VCFRowID      Allele
#>                  <Rle> <IRanges>  <Rle> | <integer> <character>
#>   rs199764262       15  48413269      * |         1           C
#>   rs199764262       15  48413269      * |         1           C
#>   rs199764262       15  48413269      * |         1           C
#>   rs199764262       15  48413269      * |         1           C
#>                         Consequence      IMPACT      SYMBOL            Gene
#>                         <character> <character> <character>     <character>
#>   rs199764262      missense_variant    MODERATE     SLC24A5 ENSG00000188467
#>   rs199764262      missense_variant    MODERATE     SLC24A5 ENSG00000188467
#>   rs199764262 upstream_gene_variant    MODIFIER     SLC24A5 ENSG00000188467
#>   rs199764262      missense_variant    MODERATE     SLC24A5 ENSG00000188467
#>               Feature_type         Feature         BIOTYPE        EXON
#>                <character>     <character>     <character> <character>
#>   rs199764262   Transcript ENST00000449382  protein_coding         1/8
#>   rs199764262   Transcript ENST00000482911  protein_coding         1/2
#>   rs199764262   Transcript ENST00000463289 retained_intron        <NA>
#>   rs199764262   Transcript ENST00000341459  protein_coding         1/9
#>                    INTRON                  HGVSc                  HGVSp
#>               <character>            <character>            <character>
#>   rs199764262        <NA> ENST00000449382.2:c... ENSP00000389966.2:p...
#>   rs199764262        <NA> ENST00000482911.2:c... ENSP00000453395.1:p...
#>   rs199764262        <NA>                   <NA>                   <NA>
#>   rs199764262        <NA> ENST00000341459.3:c... ENSP00000341550.3:p...
#>               cDNA_position CDS_position Protein_position Amino_acids
#>                 <character>  <character>      <character> <character>
#>   rs199764262            49           28               10         A/P
#>   rs199764262            69           28               10         A/P
#>   rs199764262          <NA>         <NA>             <NA>        <NA>
#>   rs199764262           101           28               10         A/P
#>                    Codons Existing_variation    DISTANCE      STRAND
#>               <character>        <character> <character> <character>
#>   rs199764262     Gcg/Ccg        rs199764262        <NA>           1
#>   rs199764262     Gcg/Ccg        rs199764262        <NA>           1
#>   rs199764262        <NA>        rs199764262         904           1
#>   rs199764262     Gcg/Ccg        rs199764262        <NA>           1
#>                     FLAGS VARIANT_CLASS SYMBOL_SOURCE     HGNC_ID   CANONICAL
#>               <character>   <character>   <character> <character> <character>
#>   rs199764262        <NA>           SNV          HGNC       20611        <NA>
#>   rs199764262        <NA>           SNV          HGNC       20611        <NA>
#>   rs199764262        <NA>           SNV          HGNC       20611        <NA>
#>   rs199764262        <NA>           SNV          HGNC       20611         YES
#>                       TSL      APPRIS        CCDS            ENSP   SWISSPROT
#>               <character> <character> <character>     <character> <character>
#>   rs199764262        <NA>        <NA>        <NA> ENSP00000389966 NCKX5_HUMAN
#>   rs199764262        <NA>        <NA>        <NA> ENSP00000453395        <NA>
#>   rs199764262        <NA>        <NA>        <NA>            <NA>        <NA>
#>   rs199764262        <NA>        <NA> CCDS10128.1 ENSP00000341550 NCKX5_HUMAN
#>                     TREMBL       UNIPARC  GENE_PHENO                   SIFT
#>                <character>   <character> <character>            <character>
#>   rs199764262         <NA> UPI000156577B           1 deleterious_low_conf..
#>   rs199764262 H0YLZ0_HUMAN UPI000153D833           1 deleterious_low_conf..
#>   rs199764262         <NA>          <NA>           1                   <NA>
#>   rs199764262         <NA> UPI0000242BC9           1 tolerated_low_confid..
#>                             PolyPhen                DOMAINS HGVS_OFFSET
#>                          <character>            <character> <character>
#>   rs199764262          benign(0.037) Cleavage_site_(Signa..        <NA>
#>   rs199764262 possibly_damaging(0... Cleavage_site_(Signa..        <NA>
#>   rs199764262                   <NA>                   <NA>        <NA>
#>   rs199764262          benign(0.177) hmmpanther:PTHR10846..        <NA>
#>                      GMAF     AFR_MAF     AMR_MAF     EAS_MAF     EUR_MAF
#>               <character> <character> <character> <character> <character>
#>   rs199764262    C:0.0002    C:0.0008         C:0         C:0         C:0
#>   rs199764262    C:0.0002    C:0.0008         C:0         C:0         C:0
#>   rs199764262    C:0.0002    C:0.0008         C:0         C:0         C:0
#>   rs199764262    C:0.0002    C:0.0008         C:0         C:0         C:0
#>                   SAS_MAF      AA_MAF      EA_MAF    ExAC_MAF ExAC_Adj_MAF
#>               <character> <character> <character> <character>  <character>
#>   rs199764262         C:0    C:0.0014         C:0 C:1.647e-04   C:0.000169
#>   rs199764262         C:0    C:0.0014         C:0 C:1.647e-04   C:0.000169
#>   rs199764262         C:0    C:0.0014         C:0 C:1.647e-04   C:0.000169
#>   rs199764262         C:0    C:0.0014         C:0 C:1.647e-04   C:0.000169
#>               ExAC_AFR_MAF ExAC_AMR_MAF ExAC_EAS_MAF ExAC_FIN_MAF ExAC_NFE_MAF
#>                <character>  <character>  <character>  <character>  <character>
#>   rs199764262   C:0.001897  C:8.803e-05          C:0          C:0          C:0
#>   rs199764262   C:0.001897  C:8.803e-05          C:0          C:0          C:0
#>   rs199764262   C:0.001897  C:8.803e-05          C:0          C:0          C:0
#>   rs199764262   C:0.001897  C:8.803e-05          C:0          C:0          C:0
#>               ExAC_OTH_MAF ExAC_SAS_MAF    CLIN_SIG     SOMATIC       PHENO
#>                <character>  <character> <character> <character> <character>
#>   rs199764262          C:0          C:0        <NA>        <NA>        <NA>
#>   rs199764262          C:0          C:0        <NA>        <NA>        <NA>
#>   rs199764262          C:0          C:0        <NA>        <NA>        <NA>
#>   rs199764262          C:0          C:0        <NA>        <NA>        <NA>
#>                    PUBMED  MOTIF_NAME   MOTIF_POS HIGH_INF_POS
#>               <character> <character> <character>  <character>
#>   rs199764262        <NA>        <NA>        <NA>         <NA>
#>   rs199764262        <NA>        <NA>        <NA>         <NA>
#>   rs199764262        <NA>        <NA>        <NA>         <NA>
#>   rs199764262        <NA>        <NA>        <NA>         <NA>
#>               MOTIF_SCORE_CHANGE  CADD_PHRED    CADD_RAW
#>                      <character> <character> <character>
#>   rs199764262               <NA>       12.78    1.398183
#>   rs199764262               <NA>       12.78    1.398183
#>   rs199764262               <NA>       12.78    1.398183
#>   rs199764262               <NA>       12.78    1.398183
#>   -------
#>   seqinfo: 86 sequences from "b37" genome