The Shiny Variant Explorer (tSVE) web-application

Currently unsupported — Package undergoing major updates.

This function starts the interactive tSVE shiny web-application that allows to interactively load and visualise genetic variants and their Ensembl Variant Effect Predictor (VEP) predictions using the package methods. All arguments after the ... set default values for the application (e.g. widgets).

tSVE(
    ...,
    refGT = "0|0",
    hetGT = c("0|1", "1|2", "0|2", "1|0", "2|1", "2|0"),
    altGT = c("1|1", "2|2"),
    vepKey = "CSQ",
    refSuffix = "REF", hetSuffix = "HET", altSuffix = "ALT",
    aafSuffix = "AAF", mafSuffix = "MAF",
    genoHeatmap.height = "500px",
    options.width = 120,
    autodetectGTimport = FALSE
)

Arguments

...

Additional arguments passed to the runApp function from the shiny package.

refGT

Default homozygote reference genotypes.

hetGT

Default heterozygote genotypes.

altGT

Default homozygote alternate genotypes.

vepKey

Default INFO key for the VEP prediction field.

refSuffix

Default INFO key suffix used to store the data for homozygote reference genotypes.

hetSuffix

Default INFO key suffix used to store the data for heterozygote genotypes.

altSuffix

Default INFO key suffix used to store the data for homozygote alternate genotypes.

aafSuffix

Default INFO key suffix used to store the data for alternate allele frequency.

mafSuffix

Default INFO key suffix used to store the data for minor allele frequency.

genoHeatmap.height

Default height (in pixels) of the heatmap that represents the genotype of each variant in each sample.

options.width

Sets options("width").

autodetectGTimport

Default checkbox value. If FALSE, genotypes (ref, het, alt) are taken as is from the Advanced settings panel. If TRUE, genotypes selected in the Advanced settings panel are updated using the autodetectGenotypes method, immediately after variants are imported.

References

Interface to EnsDb adapted from the ensembldb package.

Value

Not applicable (yet).

Author

Kevin Rue-Albrecht

See also

Examples

if (interactive()){
    runEnsDbApp()
}