VEP predictions of variants observed in samples

Returns VEP predictions for variants observed (uniquey) in samples associated with a given phenotype level.

# S4 method for class 'ExpandedVCF'
vepInPhenoLevel(
    vcf, phenoCol, level, vepCol, unique = FALSE)

Arguments

vcf

ExpandedVCF object.

metadata(vcf)[["TVTBparam"]] must contain a TVTBparam object.

phenoCol

Name of a column in pheno.

level

Phenotype level; only variants observed in at least one sample will be considered.

vepCol

VEP prediction fields; character vector of metadata columns in parseCSQToGRanges(vcf).

unique

If TRUE, consider only variants unique to the phenotype level (i.e. absent from all other phenotype levels).

Value

A GRanges including all VEP predictions associated with a variant seen in at least one sample (heterozygote or alternate homozygote) associated with the phenotype level. The GRanges contains at least one column for the VEP prediction value. Additional columns containing another VEP prediction field may be added using the facet argument.

Note

If available, "Feature" is a recommended value for this argument, as VEP typically produce one prediction per variant per feature.

Warning

A warning message is issued if genotypes are not fully defined in the TVTBparam.

Author

Kevin Rue-Albrecht

See also

Examples

# Example data ----

# VCF file
vcfFile <- system.file("extdata", "moderate.vcf", package = "TVTB")

# Phenotype file
phenoFile <- system.file("extdata", "moderate_pheno.txt", package = "TVTB")
phenotypes <- S4Vectors::DataFrame(
    read.table(file = phenoFile, header = TRUE, row.names = 1))

# TVTB parameters
tparam <- TVTBparam(Genotypes("0|0", c("0|1", "1|0"), "1|1"))

# Pre-process variants
vcf <- VariantAnnotation::readVcf(
    vcfFile, param = tparam, colData = phenotypes)
vcf <- VariantAnnotation::expand(vcf, row.names = TRUE)

# Example usage ----

vepInPhenoLevel(vcf, "super_pop", "AFR", c("CADD_PHRED", "Feature", "IMPACT"))
#> GRanges object with 48 ranges and 3 metadata columns:
#>               seqnames    ranges strand |  CADD_PHRED         Feature
#>                  <Rle> <IRanges>  <Rle> | <character>     <character>
#>   rs199764262       15  48413269      * |       12.78 ENST00000449382
#>   rs199764262       15  48413269      * |       12.78 ENST00000482911
#>   rs199764262       15  48413269      * |       12.78 ENST00000463289
#>   rs199764262       15  48413269      * |       12.78 ENST00000341459
#>     rs1426654       15  48426484      * |       11.03 ENST00000463289
#>           ...      ...       ...    ... .         ...             ...
#>   rs146726548       15  48434487      * |        25.6 ENST00000267836
#>   rs146726548       15  48434487      * |        25.6 ENST00000341459
#>   rs146726548       15  48434487      * |        25.6 ENST00000560172
#>   rs146726548       15  48434487      * |        25.6 ENST00000324324
#>   rs146726548       15  48434487      * |        25.6 ENST00000449382
#>                    IMPACT
#>               <character>
#>   rs199764262    MODERATE
#>   rs199764262    MODERATE
#>   rs199764262    MODIFIER
#>   rs199764262    MODERATE
#>     rs1426654    MODIFIER
#>           ...         ...
#>   rs146726548    MODIFIER
#>   rs146726548    MODERATE
#>   rs146726548    MODIFIER
#>   rs146726548    MODIFIER
#>   rs146726548    MODERATE
#>   -------
#>   seqinfo: 86 sequences from "b37" genome